Restructure mc-common lib.rs
